move autoscaling to a deprecated command
