+++ helpgen/flyctlhelp.toml
@@ -510,13 +510,13 @@ longHelp  = """List all the releases of the application onto the Fly platform,
 including type, when, success/fail and which user triggered the release.
 """
 
-[scale]
-usage     = "scale"
-shortHelp = "Scale App resources"
-longHelp  = """Scale application resources
+[autoscale]
+usage     = "autoscale"
+shortHelp = "Autoscaling App resources"
+longHelp  = """Autoscaling application resources
 """
 
-    [scale.balanced]
+    [autoscale.balanced]
     usage     = "balanced"
     shortHelp = "Configure a traffic balanced App with params (min=int max=int)"
     longHelp  = """Configure the App to balance regions based on traffic with given parameters:
@@ -525,7 +525,7 @@ min=int - minimum number of instances to be allocated from region pool.
 max=int - maximum number of instances to be allocated from region pool.
 """
 
-    [scale.standard]
+    [autoscale.standard]
     usage     = "standard"
     shortHelp = "Configure a standard balanced App with params (min=int max=int)"
     longHelp  = """Configure the App without traffic balancing with the given parameters:
@@ -534,19 +534,25 @@ min=int - minimum number of instances to be allocated from region pool.
 max=int - maximum number of instances to be allocated from region pool.
 """
 
-    [scale.show]
+    [autoscale.show]
     usage     = "show"
-    shortHelp = "Show current scaling configuration"
-    longHelp  = """Show current scaling configuration
+    shortHelp = "Show current autoscaling configuration"
+    longHelp  = """Show current autoscaling configuration
 """
 
-    [scale.set]
+    [autoscale.set]
     usage     = "set"
-    shortHelp = "Set current models scaling parameters"
-    longHelp  = """Allows the setting of the current models scaling parameters:
+    shortHelp = "Set current models autoscaling parameters"
+    longHelp  = """Allows the setting of the current models autoscaling parameters:
 
 min=int - minimum number of instances to be allocated from region pool. 
 max=int - maximum number of instances to be allocated from region pool.
+"""
+
+[scale]
+usage     = "scale"
+shortHelp = "Scale App resources"
+longHelp  = """Scale application resources
 """
 
     [scale.vm]
@@ -577,6 +583,11 @@ For pricing, see https://fly.io/docs/about/pricing/
 For pricing, see https://fly.io/docs/about/pricing/
 """
 
+    [scale.show]
+    usage     = "show"
+    shortHelp = "Show current resources"
+    longHelp  = """Show current resources
+"""
 
 [secrets]
 usage     = "secrets"
